+++ sql/emaj--3.0.0--3.1.0.sql
@@ -401,6 +401,9 @@ SELECT CASE WHEN EXISTS (SELECT 1 FROM emaj.emaj_rlbk)
               THEN setval('emaj.emaj_rlbk_rlbk_id_seq', (SELECT max(rlbk_id) FROM emaj.emaj_rlbk))
        END;
 
+-- and finaly drop the temporary table
+DROP TABLE emaj_rlbk_old;
+
 --
 -- change the emaj_rlbk_plan table structure
 --