+++ CHANGES.md
@@ -6,8 +6,8 @@ E-Maj - Change log
 
 
 ###Bug fixes:###
-
-
+  * Fix some missing temporary table drops in several old extension upgrade
+    scripts.
 
 3.3.0 (2020-Mar-14)
 ------